+++ opyprint/logger/__init__.py
@@ -1,9 +1,11 @@
+from .logged_mixin import LoggedMixin
 from .logger import Logger
 from .print_logger import PrintLogger
 from .void_logger import VoidLogger
 
 __all__ = [
-    'Logger',
-    'PrintLogger',
-    'VoidLogger',
+    "LoggedMixin",
+    "Logger",
+    "PrintLogger",
+    "VoidLogger",
 ]